requirements.txt
addict==2.4.0
deepspeed==0.16.4
fiftyone==1.3.1
gradio==5.20.1
huggingface_hub==0.28.1
ipdb==0.13.13
llava.egg==info
matplotlib==3.10.1
numpy==2.2.3
onnxruntime==1.21.0
opencv_python==4.11.0.86
Pillow==11.1.0
pycocotools==2.0.8
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
safetensors==0.5.3
setuptools==75.8.0
spacy==3.8.3
supervision==0.25.1
tabulate==0.9.0
termcolor==2.5.0
timm==1.0.15
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.38.2
typer==0.15.2
yapf==0.43.0
tabulate==0.9.0